AUJM1-53: Remove spring scanner dependency and refactor spring config files
