+++ pom.xml
@@ -30,7 +30,6 @@
         <amps.version>8.7.0</amps.version>
         <osgi.javaconfig.version>0.5.0</osgi.javaconfig.version>
         <spring.version>5.2.21.RELEASE</spring.version>
-        <atlassian.spring.scanner.version>2.2.4</atlassian.spring.scanner.version>
 
         <java.version>1.8</java.version>
         <kotlin.version>1.7.20</kotlin.version>
@@ -110,19 +109,6 @@
             <groupId>com.atlassian.plugins</groupId>
             <artifactId>atlassian-plugins-osgi-javaconfig</artifactId>
             <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.atlassian.plugin</groupId>
-            <artifactId>atlassian-spring-scanner-annotation</artifactId>
-            <version>${atlassian.spring.scanner.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.atlassian.plugin</groupId>
-            <artifactId>atlassian-spring-scanner-runtime</artifactId>
-            <version>${atlassian.spring.scanner.version}</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
@@ -296,22 +282,6 @@
                         <Spring-Context>*</Spring-Context>
                         <Atlassian-Scan-Folders>META-INF/plugin-descriptors</Atlassian-Scan-Folders>
                     </instructions>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>com.atlassian.plugin</groupId>
-                <artifactId>atlassian-spring-scanner-maven-plugin</artifactId>
-                <version>${atlassian.spring.scanner.version}</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>atlassian-spring-scanner</goal>
-                        </goals>
-                        <phase>process-classes</phase>
-                    </execution>
-                </executions>
-                <configuration>
-                    <verbose>false</verbose>
                 </configuration>
             </plugin>
         </plugins>
